device-tree: Fix issues relevant to the ps-pl resets and clock node order
The PL IPs must be held in reset during the clock switch over but in the
current flow, we are doing the PS-PL reset before configuring the clocks
and there is no reset handling after the clock change. So, we can’t guarantee
the functionality if we ignore reset during the clock switchover. The working
order is to enable the clocks and then do the reset. To address this issue,
update the AFI node with ps-pl rests and move the clock nodes to the top of
afi node.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>