common: Simplify the logic that ensures no duplication in the reg property
To avoid duplication of addresses within reg property, there is a legacy
condition in SDT repo (7e0fd7c255c9 : Update the ddr4 driver if it has more memory
ranges). This condition is not capable of working with the 40 bit addresses and
needs an update. Simplify this whole logic by using the list matching with
$base and $high instead of string matching of reg with $base $size. "reg" can be
of variable size (either 1 or 2 addr cell and size cell) which makes it
inapproapriate for matching. Before generating reg from the base and size, save
them in a list and check if the values are already present in the list before
proceeding in the next iteration.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>