common: Introduce a new flag for the designs having more than 32 bit MBs
When the design has a SOFT MB that can support more than 32-bit, the reg property
for all the memory nodes and the peripheral nodes mapped to that MB should be
having 4 cells regardless of the starting address of these memory IPs. Add the
support for the same.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>