device_tree: Avoid generating the reference node for sata when it is not mapped to any proc
There can be ZynqMP designs where psu_sata is part of the hsi get_cells -hier
output but not in the hsi get_mem_ranges output which means that it is not mapped
to any proc. Status "okay" is being added to SATA nodes in such cases as well.
It is leading to the deletion of this node in Linux DT as the node is not mapped
in APU address-map and the status is "okay". This deletion leads to failure in
applying static overlays that has sata in it. Therefore, avoid generating the
reference node and the common properties for sata when it is not mapped to any
proc. This will keep the sata node from zynqmp.dtsi as is in the "disabled" state.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>