device_tree: Remove unneeded proc gen_ctrl_compatible
The proc gen_ctrl_compatible was created earlier to capture
info about control registers in SDT but as these are not being
used in the SW stack, this proc is not needed anymore and hence
remove this proc. Having this proc unnecessarily is causing other
issues where the PL IPs that are mapped to a processor conflicts with
these addresses.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>

Acked-for-series: Onkar Harsh <onkar.harsh@amd.com>