sdt: Fix logic in generating memory nodes
Generate memory nodes of DDR and linear spi only if these IPs
are mapped to any of the processors in the design. Generating
memory node even when these arent mapped to any of the processor
would result in invalid reg propety and having an invalid reg values
will result in multiple issues in the downstream tools.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>