device_tree: Re-order PR region, AFI and Clock nodes
Some of the PL IPs present in the design may not have drivers and some PL
drivers may not clock adopted. Without enabling the relevant PS-PL clock,
attempts by drivers or users to access the PL IP address space can lead
to system hangs or crashes. To avoid this issue, move the required nodes
(PR region, AFI, and PS-PL clock) to the top of the sequence to enable the
relevant PS-PL clocks before probing the PL drivers.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>