ddrps: Add filter to check MEM_TYPE for processing ddrps IPs
old mig series IPs (ddr3, ddr4 etc.) show the controller and the memory address
blocks both in their get_mem_ranges objects. This has not been the case with the
normal ddr IPs, they only show the memory addresses. The MEM_TYPE needs to be
checked for all the IPs getting serviced by ddrps.tcl to avoid the controller
addresses. This was being used in the memory_tests.tcl in embeddedsw.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>